fix firefox bug, add column separators
There was a bug where mouseover on the titles caused a reflow of the
column text in Firefox. The link icon now uses opactiy instead of
display to be hidden, and I also added a tiny transition effect.

Also added a column separator to make text more legible when using
columns.